src/requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
feedparser==5.2.1
html2text==2019.9.26
idna==2.8
psycopg2-binary==2.8.4
pymongo==3.10.0
requests==2.22.0
requests-file==1.4.3
six==1.13.0
soupsieve==1.9.5
tldextract==2.2.2
urllib3==1.25.7
